+++ metadata.rb
@@ -1,4 +1,4 @@
-name 'al_agent'
+name 'al_agents'
 maintainer 'John Ramos DualSpark'
 maintainer_email 'john.ramos@dualspark.com'
 source_url 'https://github.com/alertlogic/al_agents'
@@ -6,7 +6,7 @@ issues_url 'https://github.com/alertlogic/al_agents/issues'
 license 'All rights reserved'
 description 'Installs/Configures the Alert Logic Agent'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version '1.0.1'
+version '1.0.2'
 
 depends 'selinux_policy'
 depends 'rsyslog'