+++ CHANGELOG.md
@@ -1,11 +1,19 @@
-al_agent CHANGELOG
+al_agents CHANGELOG
 ==================
 
-This file is used to list changes made in each version of the al_agent cookbook.
+This file is used to list changes made in each version of the al_agents cookbook.
+
+0.1.2
+-----
+- John Ramos - changed the name of the cookbook from al_agent to al_agents
+
+0.1.1
+-----
+- John Ramos - Fixed source and image urls
 
 0.1.0
 -----
-- [your_name] - Initial release of al_agent
+- John Ramos - Initial release of al_agents
 
 - - -
 Check the [Markdown Syntax Guide](http://daringfireball.net/projects/markdown/syntax) for help with Markdown.